refactor: remove shader_parsing helper, add parser class
